requirements.txt
uvicorn
fastapi
langchain
langchain-community
openai
pytest
ipython
python-dotenv
jinja2
nest_asyncio
